# openapi.yml
openapi: 3.0.3
info:
  title: Simple Social API
  version: "1.0.0"
  description: A simple API for posts and comments
servers:
  - url: http://localhost:8000

paths:
  /posts/:
    post:
      summary: Create a new post
      operationId: create_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '200':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostRead'
    get:
      summary: Get all posts
      operationId: get_all_posts
      parameters:
        - name: text
          in: query
          required: false
          description: Search term for text
          schema:
            type: string
        - name: user
          in: query
          required: false
          description: Filter by author username
          schema:
            type: string
      responses:
        '200':
          description: A list of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostRead'
        '404':
          description: No posts found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts/{post_id}:
    get:
      summary: Get a post by ID
      operationId: get_post_by_id
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostRead'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /posts/{post_id}/comments:
    get:
      summary: Get comments for a post
      operationId: get_comments_for_post_id
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
        - name: text
          in: query
          required: false
          description: Search term in comment text
          schema:
            type: string
        - name: user
          in: query
          required: false
          description: Filter by comment user
          schema:
            type: string
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommentRead'
        '404':
          description: No comments found for this post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a comment for a post
      operationId: create_comment_for_post
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentCreate'
      responses:
        '200':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentRead'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /comments/{comment_id}:
    get:
      summary: Get a comment by its ID
      operationId: get_comment_by_id
      parameters:
        - name: comment_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Comment found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentRead'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:

    PostCreate:
      type: object
      required:
        - image
        - text
        - user
      properties:
        image:
          type: string
          description: >
            Base64-encoded PNG or a data URL (e.g. "data:image/png;base64,...").
          example: "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgA..."
        text:
          type: string
          example: Hello world
        user:
          type: string
          example: alice

    PostRead:
      allOf:
        - $ref: '#/components/schemas/PostCreate'
        - type: object
          required:
            - id
            - created_at
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time

    CommentCreate:
      type: object
      required:
        - text
        - user
      properties:
        text:
          type: string
          example: nice post!
        user:
          type: string
          example: alicia

    CommentRead:
      allOf:
        - $ref: '#/components/schemas/CommentCreate'
        - type: object
          required:
            - id
            - super_id
            - created_at
          properties:
            id:
              type: integer
            super_id:
              type: integer
            created_at:
              type: string
              format: date-time

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string